core: fix blocklist not updated
It is possible that addBlackListFromUrl() takes too much time fetching blacklist from the web and isBlocked() be triggered before addBlackListFromUrl() completes, and causes _blacklist be set and not updated after addBlackListFromUrl() completes.

The new algorithm fixes the issue by introducing a new _listUpdated variable and updates _blacklist and _whitelist if there has been an update.